feat(game,data): рубль — деньги-предмет; count u16, стак до 65535 (сейв v14)
Инкремент A бартер-экономики, фундамент. Решение владельца: «стак честно,
u16» — потолок 65535 (2^16 в u16 не влезает ровно на единицу; ровная
степень двойки была бы 32768 и потеряла бы половину — взят max u16).

Ширина: ItemSlot.count u8 -> u16 (клетка 4 -> 6 Б, явный pad; сетка 384 Б),
Container::count, Pickup.count, KindDrop.count — u16; ItemDef.stackMax u16,
перекладка полей держит строку в 28 Б без внутреннего паддинга. Закон
сложения не изменился: сумма в int, кламп к stackMax, потом store — все
явные u8-касты на путях счёта выметены (они бы резали молча, компилятор
такое не ловит). Пул: inv_ 256 -> 384 Б/строку, резидент после init()
306 -> 434 MiB (мера в suite_npcpool переколота на сегодняшние ширины).

Рубль: строка ruble в items.csv (443; value=1, stack 65535, монета 1 г,
craft_metal=1 — переплавка, пин 5589 -> 5590), kItemRuble эмитится
генератором и обязателен — таблица без денег не собирается. Касса в
контейнерах: последний слот, PublicBox 0..120 руб флэт на любой глубине
(строка экономики референса), RoomStash/Safe 0..своей доли капа банды,
WeaponCrate без денег; пул предметов пуст — касса всё равно едет
(коридорный PublicBox). Сдвиг id >= 343: kIt* в loot_table.cpp и
тест-локальные id перевыведены из CSV (чек-сумма поймала, как и обещала).

Сейв v14: слот на проводе 5 Б (item u16 + count u16 + condition u8),
kInventoryWire 256 -> 320, kSaveFixedWire 931 -> 995; пул-блоб тем же
штрихом (kPoolRowWire 385). v13 отвергается дважды: версией и
фингерпринтом каталога.

ctest 6/6 (game_test 242646/0, пин переколот с историей); живой прогон
--shot бутится, толпа ходит. Дальше блок 2: сейв v15 — полная
персистенция ящиков и трупов через перезаход этажа.

+++ src/app/main.cpp
@@ -5744,7 +5744,7 @@ int main(int argc, char** argv) {
                         game::item_def(boxC->item[i]).value * moved;
                     loot += v;
                     containerTake += v;
-                    boxC->count[i] = static_cast<std::uint8_t>(unplaced);
+                    boxC->count[i] = unplaced;
                     if (boxC->count[i] == 0) boxC->item[i] = game::kInvalidItem;
                 };
                 auto take_corpse_slot = [&](int i) {
@@ -5757,7 +5757,7 @@ int main(int argc, char** argv) {
                         static_cast<std::uint16_t>(s.count - unplaced);
                     if (moved == 0) return;
                     loot += game::item_def(s.item).value * moved;
-                    s.count = static_cast<std::uint8_t>(unplaced);
+                    s.count = unplaced;
                     if (s.count == 0) s = game::ItemSlot{};
                 };
                 // Пустая цель помечается ПОСЛЕ мутаций: opened — память карты
@@ -5816,7 +5816,7 @@ int main(int argc, char** argv) {
                             // Сначала доложить в одноимённый стек ТОГО ЖЕ
                             // износа (закон inventory_give: смешать байты —
                             // заразить свежий стек), потом в пустую ячейку.
-                            const std::uint8_t stackMax =
+                            const std::uint16_t stackMax =
                                 game::item_def(s.item).stackMax;
                             for (int i = 0; i < game::kContainerSlots && !placed;
                                  ++i)
@@ -5836,7 +5836,7 @@ int main(int argc, char** argv) {
                                     placed = true;
                                 }
                         } else if (corpseC) {
-                            const std::uint8_t stackMax =
+                            const std::uint16_t stackMax =
                                 game::item_def(s.item).stackMax;
                             for (std::size_t i = 0;
                                  i < game::kMaxCorpseSlots && !placed; ++i) {
